Get the broker working with osb interfaces
* Commit actually makes the broker a HydroBroker. A lot of work getting
the broker to use osb types, and conversions to the apb world once they
hit the apb layer. There's a lot of duplication between types, but I
tried to not change too much of what already existed. It makes sense to
dry things up in an additional commit on top of this and see if we
can do a better job with the casts.